refactor: change html-main tag
