+++ src/assets/scripts/types/Interfaces.ts
