+++ src/assets/styles/_index.scss
@@ -8,8 +8,11 @@
   justify-content: space-between;
 }
 
-.header{
+.header, .main, .footer {
   padding: 10px 100px;
+}
+
+.header{
   height: 10vh;
   background-color: #ffffff;
 }
@@ -52,10 +55,12 @@
 
 .main{
   flex: 1 0 auto;
+  .main__filters{
+    background-color: #041e41;
+  }
 }
 
 .footer{
-  padding: 10px 100px;
   .footer__left-items{
     display: flex;
     flex-direction: column;
