+++ src/index.html
@@ -24,7 +24,13 @@
       </div>
     </div>
   </header>
-  <main class="main"></main>
+  <main class="main">
+    <div class="container main__container">
+      <section class="main__filters">
+        <div></div>
+      </section>
+    </div>
+  </main>
   <footer class="footer">
     <div class="container footer__container">
       <div class="footer__left-items">
